[REFACTOR] Move constants file to env dir
