+++ water-level-monitoring-subsystem/src/utils/WifiConnection.cpp
@@ -1,6 +1,6 @@
 #include "WifiConnection.h"
 #include <WiFi.h>
-#include "../constants.h"
+#include "../env/constants.h"
 
 /**
  * @brief Construct a new WifiConnection object
